+++ viewer/module/timeline/timelinepanel.h
@@ -57,7 +57,6 @@ private:
     void updateMenuContents();
     void rotateImage(const QString &path, int degree);
     void onThemeChanged(ViewerThemeManager::AppTheme theme);
-    void showPrintDialog(const QStringList &paths);
 
 private:
     QMenu               *m_menu;