+++ viewer/module/timeline/timelinepanel_menu.cpp
@@ -24,6 +24,9 @@
 #include "utils/baseutils.h"
 #include "utils/imageutils.h"
 #include "widgets/dialogs/filedeletedialog.h"
+#include "widgets/printoptionspage.h"
+#include "widgets/printhelper.h"
+
 #include <QMenu>
 #include <QShortcut>
 #include <QStyleFactory>
@@ -236,7 +239,7 @@ void TimelinePanel::onMenuItemClicked(QAction *action)
         dApp->signalM->startSlideShow(vinfo);
         break;
     case IdPrint: {
-        showPrintDialog(paths);
+        PrintHelper::showPrintDialog(paths);
         break;
     }
     case IdAddToAlbum: {
@@ -307,44 +310,4 @@ void TimelinePanel::rotateImage(const QString &path, int degree)
         qDebug() << "Rotate finish!";
         m_frame->updateScrollRange();
     }
-}
-
-void  TimelinePanel::showPrintDialog(const QStringList &paths) {
-    QPrinter printer;
-    printer.setOutputFormat(QPrinter::PdfFormat);
-    QPixmap img;
-
-    QPrintDialog* printDialog = new QPrintDialog(&printer, this);
-    printDialog->resize(400, 300);
-    if (printDialog->exec() == QDialog::Accepted) {
-        QPainter painter(&printer);
-        QRectF drawRectF; QRect wRect;
-        QList<QString>::const_iterator i;
-        for(i = paths.begin(); i!= paths.end(); ++i){
-            if (!img.load(*i)) {
-                qDebug() << "img load failed" << *i;
-                continue;
-            }
-            wRect = printer.pageRect();
-            if (img.width() > wRect.width() || img.height() > wRect.height()) {
-                img = img.scaled(wRect.size(), Qt::KeepAspectRatio,
-                                 Qt::SmoothTransformation);
-            }
-            drawRectF = QRectF(qreal(wRect.width() - img.width())/2,
-                               qreal(wRect.height() - img.height())/2,
-                              img.width(), img.height());
-            painter.drawPixmap(drawRectF.x(), drawRectF.y(), img.width(),
-                               img.height(), img);
-            if (i != paths.end() - 1)
-                printer.newPage();
-        }
-        painter.end();
-        qDebug() << "print succeed!";
-        return;
-    }
-
-    QObject::connect(printDialog, &QPrintDialog::finished, printDialog,
-            &QPrintDialog::deleteLater);
-
-    qDebug() << "print failed!";
 }