+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+<a name="1.2.19"></a>
+### 1.2.19 (2018-04-11)
+
+
+#### Features
+
+* **print_dialog:**  add image settings option. ([53a8c772](53a8c772))
+
+
+
+<a name="1.2.18"></a>
+### 1.2.18 (2018-03-22)
+
+
+#### Features
+
+*   set wallpaper with dbus in flatpak ([30b00248](30b00248))
+
+#### Bug Fixes
+
+*   favorite menu and title states not synced ([2b16ceaa](2b16ceaa))
+
 <a name="1.2.17"></a>
 ## 1.2.17 (2018-03-15)